+++ tests/TaggableFileDriverTest.php
@@ -4,11 +4,11 @@
  * UserFrosting Cache (http://www.userfrosting.com)
  *
  * @link      https://github.com/userfrosting/cache
- * @copyright Copyright (c) 2013-2019 Alexander Weissman
+ * @copyright Copyright (c) 2013-2021 Alexander Weissman
  * @license   https://github.com/userfrosting/cache/blob/master/LICENSE.md (MIT License)
  */
 
-namespace UserFrosting\Cache;
+namespace UserFrosting\Cache\Tests;
 
 use Illuminate\Filesystem\Filesystem;
 use Illuminate\Support\Str;